+++ package.json
@@ -17,6 +17,7 @@
     "@multiversx/sdk-core": "^12.13.0",
     "@multiversx/sdk-network-providers": "^2.1.0",
     "@multiversx/sdk-wallet": "^4.2.0",
+    "@react-navigation/bottom-tabs": "^6.5.11",
     "@react-navigation/native": "^6.1.9",
     "@react-navigation/native-stack": "^6.9.17",
     "axios": "^1.6.1",
