+++ src/features/DApp/Dapp.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import { Text } from '@gluestack-ui/themed';
+
+const Dapp = () => {
+  return <Text>Dapp</Text>;
+};
+
+export default Dapp;
