+++ src/hooks/useSendTransactionForm.ts
@@ -3,7 +3,7 @@ import { useNavigation } from '@react-navigation/native';
 import { useForm } from 'react-hook-form';
 import { UserSecretKey } from '@multiversx/sdk-wallet/out';
 
-import { ISendTransactionForm } from '@features/SendTransaction/SendTransactionForm/SendTransactionForm';
+import { ISendTransactionForm } from '@features/Account/SendTransaction/SendTransactionForm/SendTransactionForm';
 import { brodcastTransaction } from '@services/transactionService';
 import { useTransactionStore } from '@store/transactionStore/transactionStore';
 import { useWalletStore } from '@store/walletStore/walletStore';
