+++ src/navigation/index.ts
@@ -1,1 +1,2 @@
-export { default as WalletAppNavigation } from './WalletAppNavigation';
+export { default as AccountStackNavigation } from './AccountStackNavigatiom/WalletAppNavigation';
+export { default as TabsNavigation } from './TabsNavigation/TabsNavigation';
